pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "streamlit>=1.52.2",
]
